Day 14 of 365
I created a menu with options to add a store name location then owner. When that takes place a function creates an object (based on input) and adds it to an array making an array of objects. Each input adds a store. For this project I will be adding more features